requirements.txt
Flask-3.1.1 blinker-1.9.0 certifi-2025.6.15 charset-normalizer-3.4.2 click-8.1.8 colorama-0.4.6 filelock-3.18.0 fsspec-2025.5.1 huggingface-hub-0.33.1 idna-3.10 importlib-metadata-8.7.0 itsdangerous-2.2.0 jinja2-3.1.6 markupsafe-3.0.2 mpmath-1.3.0 networkx-3.2.1 numpy-2.0.2 packaging-25.0 pyyaml-6.0.2 regex-2024.11.6 requests-2.32.4 safetensors-0.5.3 sentencepiece-0.2.0 sympy-1.14.0 tokenizers-0.21.2 torch-2.7.1 tqdm-4.67.1 transformers-4.53.0 typing-extensions-4.14.0 urllib3-2.5.0 werkzeug-3.1.3 zipp-3.23.0